user recover test complete
